refactor: extract traces fetch calls from playground ui
