+++ .changeset/huge-memes-punch.md
@@ -1,0 +1,7 @@
+---
+'@mastra/playground-ui': patch
+'mastra': patch
+'create-mastra': patch
+---
+
+move the fetch traces call to the playground instead of playground-ui
