+++ packages/cli/src/playground/src/pages/agents/agent/traces.tsx
@@ -1,11 +1,12 @@
-import { AgentTraces, useTraces } from '@mastra/playground-ui';
+import { AgentTraces } from '@mastra/playground-ui';
 import { useParams } from 'react-router';
 
 import { Skeleton } from '@/components/ui/skeleton';
 
 import { useAgent } from '@/hooks/use-agents';
+import { useTraces } from '@/domains/traces/hooks/use-traces';
 
-function AgentTracesContent() {
+function AgentTracesPage() {
   const { agentId } = useParams();
   const { agent, isLoading: isAgentLoading } = useAgent(agentId!);
   const { traces, firstCallLoading, error } = useTraces(agent?.name || '');
@@ -19,10 +20,6 @@ function AgentTracesContent() {
   }
 
   return <AgentTraces traces={traces || []} error={error} className="h-[calc(100vh-40px)]" />;
-}
-
-function AgentTracesPage() {
-  return <AgentTracesContent />;
 }
 
 export default AgentTracesPage;
