+++ packages/cli/src/playground/src/pages/workflows/workflow/legacy/traces.tsx
@@ -1,12 +1,13 @@
 import { useParams } from 'react-router';
-import { useTraces, WorkflowTraces } from '@mastra/playground-ui';
+import { WorkflowTraces } from '@mastra/playground-ui';
 
 import { Skeleton } from '@/components/ui/skeleton';
 
 import { WorkflowInformation } from '@/domains/workflows/workflow-information';
 import { useLegacyWorkflow } from '@/hooks/use-workflows';
+import { useTraces } from '@/domains/traces/hooks/use-traces';
 
-function WorkflowTracesContent() {
+function WorkflowTracesPage() {
   const { workflowId } = useParams();
   const { legacyWorkflow, isLoading: isWorkflowLoading } = useLegacyWorkflow(workflowId!);
 
@@ -27,10 +28,6 @@ function WorkflowTracesContent() {
   }
 
   return <WorkflowTraces traces={traces} error={error} />;
-}
-
-function WorkflowTracesPage() {
-  return <WorkflowTracesContent />;
 }
 
 export default WorkflowTracesPage;
