+++ packages/cli/src/playground/src/pages/workflows/workflow/traces.tsx
@@ -1,12 +1,13 @@
 import { useParams, useSearchParams } from 'react-router';
-import { useTraces, WorkflowTraces } from '@mastra/playground-ui';
+import { WorkflowTraces } from '@mastra/playground-ui';
 
 import { Skeleton } from '@/components/ui/skeleton';
 
 import { WorkflowInformation } from '@/domains/workflows/workflow-information';
 import { useWorkflow } from '@/hooks/use-workflows';
+import { useTraces } from '@/domains/traces/hooks/use-traces';
 
-function WorkflowTracesContent() {
+function WorkflowTracesPage() {
   const { workflowId } = useParams();
   const [searchParams] = useSearchParams();
   const runId = searchParams.get('runId');
@@ -30,10 +31,6 @@ function WorkflowTracesContent() {
   }
 
   return <WorkflowTraces traces={traces} error={error} runId={runId || undefined} stepName={stepName || undefined} />;
-}
-
-function WorkflowTracesPage() {
-  return <WorkflowTracesContent />;
 }
 
 export default WorkflowTracesPage;
